# tests/contract/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SentinelNet",
    "description": "校园网加密流量入侵检测系统",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/alerts": {
      "get": {
        "tags": [
          "alerts",
          "alerts"
        ],
        "summary": "查询告警列表（支持分页和级别筛选）",
        "description": "获取告警列表（从 SQLite alerts 表查询）。",
        "operationId": "get_alerts_api_alerts_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "level",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "级别筛选: high/medium/low/all",
              "default": "all",
              "title": "Level"
            },
            "description": "级别筛选: high/medium/low/all"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/alerts/stats": {
      "get": {
        "tags": [
          "alerts",
          "alerts"
        ],
        "summary": "获取告警统计",
        "description": "获取告警统计（总数、类型分布、级别分布）。",
        "operationId": "get_alert_stats_api_alerts_stats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertStatsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/models": {
      "get": {
        "tags": [
          "models",
          "models"
        ],
        "summary": "获取可用模型列表",
        "description": "获取可用模型列表 — 从 registry.json 读取。",
        "operationId": "list_models_api_models_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/models/train": {
      "post": {
        "tags": [
          "models",
          "models"
        ],
        "summary": "启动模型训练",
        "description": "启动模型训练 —— 委托 `TaskRegistry` 的 `train` 任务（T2.13）。\n\n三步保护：\n1. `confirm=true` 必填，否则 400 且不启动；\n2. 已在运行时返回 409（判据来自 TaskRegistry，不是本地状态机）；\n3. 启动前备份三个产物到 `<data_dir>/training_backup/<时间戳>/`。\n\n`epochs` 仅用于判定 quick 模式（≤3），不直接等于训练轮数 ——\n训练链路的轮数由数据集配置决定（D4：不重构训练链路内部）。",
        "operationId": "start_training_api_models_train_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/models/train/status": {
      "get": {
        "tags": [
          "models",
          "models"
        ],
        "summary": "获取训练状态",
        "description": "获取训练状态 —— 真相源是 `TaskRegistry`（T2.13）。\n\n状态映射（TaskRegistry → 本响应）：\n`idle`→`idle`、`running`→`training`、`finished`→`completed`、`failed`→`failed`。\n\n`progress` 由任务窗口 elapsed/duration 算出而非硬编码：\n训练任务在本架构里是 TIMED 任务（窗口时长 = `default_duration`，默认 120s），\n所以它是\"任务窗口进度\"而不是\"训练收敛进度\" —— 训练链路内部进度需要\n改 `model/train.py` 才能拿到，属 D4 范围外。",
        "operationId": "get_train_status_api_models_train_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainStatusResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/models/{name}": {
      "delete": {
        "tags": [
          "models",
          "models"
        ],
        "summary": "删除指定模型",
        "description": "删除指定模型 — 从 registry.json 移除并删除 run 目录。\n\n悬挂指针修复（2026-09-19）：删除后按 `model/train.py` 自身规则重算\n`latest.json` / `best.json`，并把剩余条目的 `is_best` 与 best 对齐；\n删空后指针文件一并移除。注册表写失败（500）时不动指针文件。",
        "operationId": "delete_model_api_models__name__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AlertListResponse": {
        "properties": {
          "alerts": {
            "items": {
              "$ref": "#/components/schemas/AlertResponse"
            },
            "type": "array",
            "title": "Alerts"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "default": 50
          },
          "offset": {
            "type": "integer",
            "title": "Offset",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "alerts",
          "total"
        ],
        "title": "AlertListResponse",
        "description": "告警列表响应。",
        "examples": [
          {
            "alerts": [],
            "limit": 50,
            "offset": 0,
            "total": 0
          }
        ]
      },
      "AlertResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "time": {
            "type": "string",
            "title": "Time"
          },
          "level": {
            "type": "string",
            "title": "Level"
          },
          "attack_type": {
            "type": "string",
            "title": "Attack Type"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "ml_confidence": {
            "type": "number",
            "title": "Ml Confidence",
            "default": 0.0
          }
        },
        "type": "object",
        "required": [
          "id",
          "time",
          "level",
          "attack_type",
          "message"
        ],
        "title": "AlertResponse",
        "description": "单条告警响应。",
        "examples": [
          {
            "attack_type": "ddos",
            "id": 1,
            "level": "high",
            "message": "DDoS attack detected",
            "ml_confidence": 0.95,
            "time": "2026-01-01 00:00:00"
          }
        ]
      },
      "AlertStatsResponse": {
        "properties": {
          "total_alerts": {
            "type": "integer",
            "title": "Total Alerts",
            "default": 0
          },
          "alerts_by_type": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Alerts By Type"
          },
          "alerts_by_severity": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Alerts By Severity"
          }
        },
        "type": "object",
        "title": "AlertStatsResponse",
        "description": "告警统计响应。"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MessageResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "ok"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "MessageResponse",
        "description": "通用消息响应。",
        "examples": [
          {
            "message": "操作成功",
            "status": "ok"
          }
        ]
      },
      "ModelInfoResponse": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "metrics": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metrics"
          },
          "is_best": {
            "type": "boolean",
            "title": "Is Best",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "name",
          "version",
          "created_at"
        ],
        "title": "ModelInfoResponse",
        "description": "模型信息响应。"
      },
      "ModelListResponse": {
        "properties": {
          "models": {
            "items": {
              "$ref": "#/components/schemas/ModelInfoResponse"
            },
            "type": "array",
            "title": "Models"
          }
        },
        "type": "object",
        "required": [
          "models"
        ],
        "title": "ModelListResponse",
        "description": "模型列表响应。"
      },
...
      "TrainRequest": {
        "properties": {
          "dataset": {
            "type": "string",
            "title": "Dataset",
            "description": "训练数据集标识（透传给训练链路 dataset_type）",
            "default": "auto"
          },
          "epochs": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Epochs",
            "description": "训练轮数；≤3 视为 quick 模式",
            "default": 10
          },
          "confirm": {
            "type": "boolean",
            "title": "Confirm",
            "description": "**必填确认位**。训练会覆盖 model.pkl / evaluation_report.txt / confusion_matrix.png，未传 `confirm=true` 时请求被拒绝（400），不会启动训练。启动前会自动备份这三个产物到 `<data_dir>/training_backup/<时间戳>/`。",
            "default": false
          }
        },
        "type": "object",
        "title": "TrainRequest",
        "description": "训练请求（T2.13 起带破坏性操作保护）。"
      },
      "TrainStatusResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "description": "idle | training | completed | failed"
          },
          "progress": {
            "type": "number",
            "title": "Progress",
            "description": "进度 0.0–1.0，由任务窗口已耗时/总时长算出（非硬编码）",
            "default": 0.0
          },
          "epoch": {
            "type": "integer",
            "title": "Epoch",
            "default": 0
          },
          "total_epochs": {
            "type": "integer",
            "title": "Total Epochs",
            "default": 0
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "elapsed_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Elapsed Seconds",
            "description": "任务已运行秒数（来自 TaskRegistry）"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds",
            "description": "任务窗口时长（来自 TaskRegistry）"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "TrainStatusResponse",
        "description": "训练状态响应。\n\n`status` 的真相源是 `TaskRegistry` 里 `train` 任务的状态\n（T2.13 收敛；此前是 `api/models.py` 的模块级状态机，与任务编排互不感知）。",
        "examples": [
          {
            "epoch": 5,
            "progress": 0.5,
            "status": "training",
            "total_epochs": 10
          }
        ]
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}